facade admin divisions test and logic

+++ src/main/java/com/github/voragoth/drugstores/service/AdminDivisionsProviderService.java
@@ -15,7 +15,7 @@ public interface AdminDivisionsProviderService {
      * @param region la region para filtrar las comunas.
      * @return la lista comuna-llave de comunas pertenecientes a la region.
      */
-    Map<String, Integer> getCommunes(Integer region);
+    Map<String, Integer> getCommunes(Byte region);
 
     /**
      * Metodo que retorna las regiones.
