+++ src/main/java/com/github/voragoth/drugstores/service/impl/AdminDivisionsProviderServiceImpl.java
@@ -17,7 +17,7 @@ public class AdminDivisionsProviderServiceImpl implements AdminDivisionsProvider
      * {@inheritDoc}
      */
     @Override
-    public Map<String, Integer> getCommunes(Integer region) {
+    public Map<String, Integer> getCommunes(Byte region) {
         throw new IllegalStateException("metodo aun no implementado");
     }
